Nippo Editor のコンポーネントを使えるように修正
