+++ src/app/_components/uiParts/Editor/index.ts
@@ -1,0 +1,1 @@
+export { Editor } from './Editor';
