+++ src/app/_components/uiParts/Editor/styles.scss
